Refactoring swagger documents

+++ src/web/ctl/c_demo.go
@@ -18,12 +18,10 @@ type Demo struct {
 }
 
 // Query 查询数据
-// @Title 查询数据
-// @Description 查询示例数据(支持分页查询)
-// @Consumes json
+// @Summary 查询数据(分页查询)
 // @Produces json
 // @Param access-token header string false "访问令牌"
-// @Param type query string true "查询类型(分页查询：page)"
+// @Param type query string true "查询类型(分页查询：page)" page
 // @Param current query int true "分页索引" 1
 // @Param pageSize query int true "分页大小" 10
 // @Param code query string false "编号"
@@ -31,6 +29,7 @@ type Demo struct {
 // @Param status query int false "状态"
 // @Success 200 []schema.Demo "分页查询结果：{list:列表数据,pagination:{current:页索引,pageSize:页大小}}"
 // @Failure 400 string "{error:{code:0,message:未知的查询类型}}"
+// @Failure 401 string "{error:{code:0,message:未授权}}"
 // @Failure 500 string "{error:{code:0,message:服务器错误}}"
 // @Router GET /api/v1/demos
 func (a *Demo) Query(ctx *context.Context) {
@@ -61,6 +60,15 @@ func (a *Demo) QueryPage(ctx *context.Context) {
 }
 
 // Get 查询指定数据
+// @Summary 查询指定数据
+// @Produces json
+// @Param access-token header string false "访问令牌"
+// @Param id path string true "记录ID"
+// @Success 200 schema.Demo
+// @Failure 401 string "{error:{code:0,message:未授权}}"
+// @Failure 404 string "{error:{code:0,message:资源不存在}}"
+// @Failure 500 string "{error:{code:0,message:服务器错误}}"
+// @Router GET /api/v1/demos/{id}
 func (a *Demo) Get(ctx *context.Context) {
 	item, err := a.DemoBll.Get(ctx.CContext(), ctx.Param("id"))
 	if err != nil {
@@ -71,6 +79,16 @@ func (a *Demo) Get(ctx *context.Context) {
 }
 
 // Create 创建数据
+// @Summary 创建数据
+// @Consumes json
+// @Produces json
+// @Param access-token header string false "访问令牌"
+// @Param body body schema.Demo true "示例数据"
+// @Success 200 string "{record_id:记录ID}"
+// @Failure 400 string "{error:{code:0,message:无效的请求参数}}"
+// @Failure 401 string "{error:{code:0,message:未授权}}"
+// @Failure 500 string "{error:{code:0,message:服务器错误}}"
+// @Router POST /api/v1/demos
 func (a *Demo) Create(ctx *context.Context) {
 	var item schema.Demo
 	if err := ctx.ParseJSON(&item); err != nil {
@@ -88,6 +106,17 @@ func (a *Demo) Create(ctx *context.Context) {
 }
 
 // Update 更新数据
+// @Summary 更新数据
+// @Consumes json
+// @Produces json
+// @Param access-token header string false "访问令牌"
+// @Param id path string true "记录ID"
+// @Param body body schema.Demo true "示例数据"
+// @Success 200 string "{status:OK}"
+// @Failure 400 string "{error:{code:0,message:无效的请求参数}}"
+// @Failure 401 string "{error:{code:0,message:未授权}}"
+// @Failure 500 string "{error:{code:0,message:服务器错误}}"
+// @Router PUT /api/v1/demos/{id}
 func (a *Demo) Update(ctx *context.Context) {
 	var item schema.Demo
 	if err := ctx.ParseJSON(&item); err != nil {
@@ -104,6 +133,14 @@ func (a *Demo) Update(ctx *context.Context) {
 }
 
 // Delete 删除数据
+// @Summary 删除数据
+// @Produces json
+// @Param access-token header string false "访问令牌"
+// @Param id path string true "记录ID"
+// @Success 200 string "{status:OK}"
+// @Failure 401 string "{error:{code:0,message:未授权}}"
+// @Failure 500 string "{error:{code:0,message:服务器错误}}"
+// @Router DELETE /api/v1/demos/{id}
 func (a *Demo) Delete(ctx *context.Context) {
 	err := a.DemoBll.Delete(ctx.CContext(), ctx.Param("id"))
 	if err != nil {
@@ -114,8 +151,21 @@ func (a *Demo) Delete(ctx *context.Context) {
 }
 
 // DeleteMany 删除多条数据
+// @Summary 删除多条数据
+// @Produces json
+// @Param access-token header string false "访问令牌"
+// @Param batch query string true "记录ID（多个以,分隔）"
+// @Success 200 string "{status:OK}"
+// @Failure 400 string "{error:{code:0,message:无效的请求参数}}"
+// @Failure 401 string "{error:{code:0,message:未授权}}"
+// @Failure 500 string "{error:{code:0,message:服务器错误}}"
+// @Router DELETE /api/v1/demos
 func (a *Demo) DeleteMany(ctx *context.Context) {
 	ids := strings.Split(ctx.Query("batch"), ",")
+	if len(ids) == 0 {
+		ctx.ResError(errors.NewBadRequestError("无效的请求参数"))
+		return
+	}
 
 	for _, id := range ids {
 		err := a.DemoBll.Delete(ctx.CContext(), id)
@@ -129,6 +179,14 @@ func (a *Demo) DeleteMany(ctx *context.Context) {
 }
 
 // Enable 启用数据
+// @Summary 启用数据
+// @Produces json
+// @Param access-token header string false "访问令牌"
+// @Param id path string true "记录ID"
+// @Success 200 string "{status:OK}"
+// @Failure 401 string "{error:{code:0,message:未授权}}"
+// @Failure 500 string "{error:{code:0,message:服务器错误}}"
+// @Router PATCH /api/v1/demos/{id}/enable
 func (a *Demo) Enable(ctx *context.Context) {
 	err := a.DemoBll.UpdateStatus(ctx.CContext(), ctx.Param("id"), 1)
 	if err != nil {
@@ -139,6 +197,14 @@ func (a *Demo) Enable(ctx *context.Context) {
 }
 
 // Disable 禁用数据
+// @Summary 禁用数据
+// @Produces json
+// @Param access-token header string false "访问令牌"
+// @Param id path string true "记录ID"
+// @Success 200 string "{status:OK}"
+// @Failure 401 string "{error:{code:0,message:未授权}}"
+// @Failure 500 string "{error:{code:0,message:服务器错误}}"
+// @Router PATCH /api/v1/demos/{id}/disable
 func (a *Demo) Disable(ctx *context.Context) {
 	err := a.DemoBll.UpdateStatus(ctx.CContext(), ctx.Param("id"), 2)
 	if err != nil {
